requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
dateutils==0.6.12
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
mongoengine-goodjson==1.1.8
pipreqs==0.4.11
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
yarg==0.1.9
